Antigrav ability semi-implimented
The ability isn't functional.

What's missing is how it interacts with the current character controller.

I need it to be able to disable the other movement forms while it's active.

I've got the logic in place for that but i don't know how to go about disabling the current movement.