Character Animations + Improved Enemy Animations
Added the vast majority of the main character's animations. Only missing "Death" "Wall Slide" and "Swim." The character's "run" "jump" "crouch walk" and "idle" animations are timed out in anim files whereas the rest are still frames that can be pulled from the "Character Anim Sprites" folder, I wasn't sure if those would need a anim file or the sprites could be swapped out via another method when needed. I have also redone the rig and animations for the HellHound enemy as the previous version of this branch seemed to have lost all of the bones.